🚧 K4FCKS task: split Hermes adapter runtime helpers
